Filters
The image is converted to grayscale then salt and pepper noise is added with probability 0.15. Average filter and median filter of various sizes are realized. After the original image is taken, Robert filters are applied. The original image is taken again and the Canny Edge Detector in OpenCV is used.